Refactor tutorial stories
