+++ src/styles/notification.css
@@ -33,12 +33,10 @@
 
   10%, 50% {
     margin-left: 0%;
-  }
-
-  60% {
     opacity: 1;
   }
-  90% {
+
+  70% {
     opacity: 0;
   }
